CASMINST-5619, CASMINST-5764: Update management node personalization and image customization procedures (#3511)
* CASMINST-5619: Reference IUF in top-level upgrade/install procedures

Update the instructions in the CSM upgrade/README.md and
install/README.md files to refer into the appropriate IUF procedures for
performing an upgrade or install using the IUF.

Rename the main procedures in the IUF docs to more accurately describe
when the procedures are used. The "Initial Install" procedure is really
for performing install _or_ upgrade of any products other than CSM. The
"Upgrade" procedure is really for performing an upgrade of CSM along
with the other products.

Add a section at the end of `install/README.md` that refers the admin to
the `install_or_upgrade_additional_products_with_iuf.md` procedure. The
IUF procedure still includes a link back to `install/README.md` for
installing CSM, but this should already be done. The reference is left
in for completeness and to align the steps with a diagram being
developed showing the overall install procedure with IUF.

In the case of an upgrade, there are a few alternatives:

* If upgrading both CSM and additional HPE Cray EX products at the same
  time, refer to `upgrade_csm_and_additional_products_with_iuf.md`.
* If upgrading only additional products, refer the admin to the
  `install_or_upgrade_additional_products_with_iuf.md`.
* If upgrading only CSM on a CSM-only system, they should just follow
  the existing CSM upgrade instructions.

To make the alternatives clear, move `upgrade/README.md` into a new
file, `upgrade/Upgrade_CSM_Only.md`, and document the three alternatives
in a new `upgrade/README.md` which just links to each procedure.

Additional documentation updates are needed dealing with node
personalization and image customization of management nodes. These
changes follow in a subsequent commit.

* CASMINST-5764: Refine management node personalization for IUF

Refine the documentation dealing with node personalization and image
customization for management nodes. In the Alice recipe release (CSM
1.4.0), the Install Upgrade Framework (IUF) is responsible for creating
the CFS configuration which is applied to management nodes, and the same
CFS configuration is used to perform image customization of the
management node images as well as post-boot node personalization.

The IUF creates the CFS configuration during the `update-cfs-config`
stage and customizes the management node images during the
`prepare-images` stage. Both stages use `sat bootprep` with an input
file delivered with the HPC CSM Software Recipe that defines the layers
in the management node CFS configuration and defines how this CFS
configuration should be applied to the base images delivered by the CSM
product to create customized worker (and optionally, master and storage)
images.

As a result of these significant changes, the documentation dealing with
management node personalization and image customization needs sigificant
updates. Here is a summary of those changes.

During CSM initial install, use `apply_csm_configuration.sh` to create a
temporary CSM-only CFS configuration and apply it to the management
nodes. The final CFS configuration for management nodes is created and
used to customize images and personalize management nodes in the IUF
procedure. The procedures to configure passwordless SSH and the root
password and SSH keys on management nodes are moved from
`Configure_Non-Compute_Nodes_with_CFS.md` into their own separate
documents and referenced from `configure_administrative_access.md`.

Each of these separate procedures can also be run in an operational
context. To apply their configuration, they now just reference a
procedure to re-run node personalization on management nodes instead of
a procedure to create a whole new CFS configuration with bootprep files.

For an upgrade of CSM with other products, the admin follows the IUF
procedure. For an upgrade of only CSM, the procedure depends on whether
the system is CSM-only or has additional products installed. Document
the two alternatives in Stage 0.3:

* The CSM-only case uses `apply_csm_configuration.sh` to update the CFS
  configuration and apply it to management nodes. Image customization is
  not required.
* If other products are installed, follow a procedure to obtain the
  bootprep files and `session_vars.yaml` file and use `sat bootprep`
  directly to create the CFS configuration that includes the new version
  of CSM and the existing versions of other products. Create custom images
  and assign those to the management nodes in BSS.

Other changes include:

* Remove `Worker_Image_Customization.md`. This is now covered either in
  the IUF docs or in-line in Stage 0.3 for upgrades and is not needed in
  CSM installs.
* Refactor `Management_Node_Image_Customization.md` to be an
  operational procedure. Management node image customization during
  upgrade is now handled in the IUF documentation.
* Create a new `Management_Node_Personalization.md` that describes
  management node personalization with CFS and includes some common
  operational procedures.
* Move all content from `Perform_NCN_Personalization.md` into
  `Management_Node_Personalization.md`.
* Remove `NCN_Node_Personalization.md`, which had outdated information
  about editing bootprep files.

Test Description:

For the procedure to upgrade only CSM on a system installed with
additional products, I tested each command on odin as I wrote it.

These procedures have not been tested exactly as written yet. The
procedures for upgrading all products with IUF have been getting tested,
but there have not been any tests for the modified CSM fresh-install
procedure or the modified procedure for CSM-only upgrades.

* Apply suggestions from code review

Fix typos, grammatical errors, and minor style issues based on suggestions from Nathan and Dean.

Co-authored-by: Nathan Rockershousen <nathan.rockershousen@hpe.com>
Co-authored-by: Dean Roe <dean.roe@hpe.com>
Signed-off-by: Ryan Haasken <77809410+haasken-hpe@users.noreply.github.com>

* CASMINST-5764: Add prereq headings to Management_Node_Personalization.md

Per suggestion from Nathan, make put the prerequisites for each of these
procedures under a separate heading.

---------

Signed-off-by: Ryan Haasken <77809410+haasken-hpe@users.noreply.github.com>
Co-authored-by: Nathan Rockershousen <nathan.rockershousen@hpe.com>
Co-authored-by: Dean Roe <dean.roe@hpe.com>