CASMINST-6142 Fix unexported variable
The `https_proxy` var needs to be exported the way it is currently used,
which we want to avoid as it can break other commands if `https_proxy`
is left set.

This merges it back into one line while keeping the code-snippets
separate for facilitating copy-paste. Using this in one line avoids the
need to `export`, while allowing `wget` to find `https_proxy`.