CASMTRIAGE-5113: Chunk up xnames in requests to CFS (#3632)
When applying CFS configurations, the CFS API can only handle a certain
number of xnames as query parameters. This is due to the buffer size
of uwsgi. The default buffer size is too small to handle more than 34
xnames per request and sometimes fewer.

This mod breaks up requests into twenty xnames or fewer. The CFS API
can handle this many.

This is a workaround for the CSM-1.3.2 release. In later releases,
uwsgi's buffer size was increased.