CASMPET-6607: Update loftsman manifest for certmanager add checks
The prior release would only catch the cray-certmanager on a loftsman install.
Update this section to ensure we install:
- cray-drydock
- cray-certmanager
- cray-certmanager-issuers

Note the first and last helm charts are "technically" not needed but we want to
be safe and include anything that might be affected by namespace removals and
ownership transferrals.

While here also add checks to ensure that the helm chart uninstalls behave as we
expect and if any helm uninstall does not occur, exit the prereqs run.

If prereqs was not ran or did not get to this section of code we will see helm
install issues with CRD ownership due to cray-certmanager-init not being removed.