CASMPET-6582 Add roll spire postgres
spire-postgres fails restart after the spire chart has been upgraded.
This causes the postgres operator to mark it as 'SyncFailed'. Rolling
the statefulset here should fix the issue.

NOTE: This section uses kubectl rollout status to wait for the pod
restarts to complete before continuing. This only works on a node with a
RollingUpdate strategy. After these nodes are rolled the operator will
have changed the strategy to OnDelete. This command will fail after this
happens.