CASMPET-6582 Add roll spire postgres (#3835)
* CASMPET-6582 Add roll spire postgres

spire-postgres fails restart after the spire chart has been upgraded.
This causes the postgres operator to mark it as 'SyncFailed'. Rolling
the statefulset here should fix the issue.

NOTE: This section uses kubectl rollout status to wait for the pod
restarts to complete before continuing. This only works on a node with a
RollingUpdate strategy. After these nodes are rolled the operator will
have changed the strategy to OnDelete. This command will fail after this
happens.

* CASMINST-6424 Move PREFLIGHT_CHECK to bottom

This check should be run last in order to validate that all the previous
steps are considered.

* CASMPET-6582 Fix incorrect timeout

* CASMPET-6582  Fix indentions