CASMPET-6612 Improve spire-postgres health check
This check was neither idempotent, nor did it wait long enough. This
updates the rolling of the spire-postgres pods to be during the spire
chart upgrade and moves the check to be just before validation. This
should both reduce the time it takes to get through this piece as well
as increasing the likihood that it will be completed before the
validation check times out.