Adhoc NCN configuration for sysctl values
A new playbook has been provided that needs to run against booted NCN nodes to provide
updated sysctl values. Included are a set of instructions that will create a new CFS
configuration and then launch an imperative CFS session that will run the associated playbook.

NOTE: this documentation specifically references the config branch that was installed on wasp,
but this version of csm-config does not contain the referenced CFS playbook in question. As
a result, the overall configuraiton will fail; this value needs to be updated to a version of
csm-config (by branch name) that contains the top level playbook that needs to be run.