CASMTRIAGE-5856 Fix `yq4` call
The `yq4` call is invalid and causing the step to fail.

In `yq` 4.33.0 the command works, but in 4.6.0 it does not. In 4.6.0
(what's installed in 1.4) an `eval` is needed as well as a `-j`.

`-j` also works in 4.33.0 (what's installed in 1.5 and 1.6), but a
deprecation notice will print.

```
Flag --tojson has been deprecated, please use -o=json instead
```