CASMINST-6597: Add instructions to update management CFS config
Replace the instructions which run a single CFS session against
management nodes (NCNs) with instructions to call a new script which
updates the CFS configuration which applies to the management nodes and
waits for the components to reach a "configured" status.

This is a more correct way to apply the updated CSM CFS configuration
layer because it ensures that the same configuration is applied if the
node is rebooted for any reason in the future.

Test Description:
Tested the execution of this script as documented on mug.