CASMINST-6679: Use explicit cray-sat container image version
Use an explicit version tag for the cray-sat container image. The 3.25.2
version is the version included in the latest release/1.5 branch of the
csm repository. We can run into problems with caching of the image on
the Kubernetes nodes if we use the `csm-latest` tag.

An alternative to this might be to use an imagePullPolicy of Always, but
I'm not sure what other impacts that might have. Additionally, using an
explicit version is what was done for the iuf container in the iufBase
template, so do the same here for consistency.

(cherry picked from commit 3a3bf591bdfc5d1e95f3bbb97da6cf505c21656d)