CASMTRIAGE-6251: Fix cray-sat image location (#4446)
The cray-sat image has moved from sat-docker to csm-docker for CSM
1.6.0. Update the cray-sat image location in the Argo workflow template
to match. Without this fix, the `update_tags.sh` script fails when
trying to update image tags in the Argo workflow templates, which
actually causes the docs-csm RPM installation to fail because the script
is executed as an RPM post-install scriptlet.