CRAYSAT-1782: Add information about viewing CSM Docs
Migrate the information from docs-sat about the various ways to view and
navigate the documentation both online and offline. Describe how to use
the navigation menus both in the Hugo-rendered HTML and in GitHub
directly.

Perform minor refactoring of `introduction/README.md` to include a
corresponding section for each of the topics listed at the beginning of
the file for consistency.