CRAYSAT-1782: Add SAT-specific command-prompt conventions
There are a couple of SAT-specific command-prompt conventions that need
to be described in the CSM documentation command prompt conventions when
migrating docs-sat content into docs-csm.

The first is just the command-prompt convention for shells created by
the `sat bash` command. This is very similar to shells in Kubernetes
pods obtained by `kubectl exec`.

The second is a command-prompt convention for shells in which a Python
virtual environment is active. This is used in documentation that
describes installation and use of `sat` on external systems.