CRAYSAT-1782: Migrate uninstall/downgrade info from docs-sat
Migrate the topics on uninstalling and downgrading SAT versions from
docs-sat to docs-csm. Now that SAT is no longer a separate product and
is instead fully included in CSM v1.6.0, it can no longer be separately
uninstalled/removed by `prodmgr`. Update the documentation to note that
the uninstall instructions are only valid for the versions of SAT
installed as a separate product.

In addition, the ability to activate other versions of `sat` with
`prodmgr activate` is no longer supported. Instead, document how to more
simply switch between versions of the `cray-sat` container image using
the `SAT_IMAGE` environment variable. Move the documentation on this
procedure which was in the FAQ on the "SAT in CSM" topic into the
separate document which discusses uninstall and downgrade.