CRAYSAT-1782: Reference SAT configuration in install procedure
Refer the reader to the appropriate SAT configuration procedures at the
appropriate time in the CSM install procedure. This should be done at
the same time as the `cray` CLI configuration and authentication.

This should not be necessary during upgrades as the authentication
tokens should remain in place, and I don't see anywhere it is re-done
for the `cray` CLI as part of the upgrade procedure.