CRAYSAT-1782: Change structure of "SAT Dependencies" topic
The markdownlint style checker complains about multiple headings with
the same name in a single file, even if they are not sibling headings.
docs-sat had limited the check on duplicate headings to only apply to
sibling headings. Since docs-csm has not limited that check to only
siblings, it was failing the style check. Replace the headings that
specified product names with just normal paragraphs that specify product
names.