CRAYSAT-1782: Migrate docs-sat content to docs-csm (#4476)
* CRAYSAT-1782: Rewrite "SAT in CSM" section

Rewrite the "SAT in CSM" section of the documentation now that SAT is
fully included in CSM, and there is no longer a separate SAT product
moving forward.

Also rename the operations folder from `sat` to `system_admin_toolkit`
for more consistency with other folder names which use unabbreviated
component names.

* CRAYSAT-1782: Add information about viewing CSM Docs

Migrate the information from docs-sat about the various ways to view and
navigate the documentation both online and offline. Describe how to use
the navigation menus both in the Hugo-rendered HTML and in GitHub
directly.

Perform minor refactoring of `introduction/README.md` to include a
corresponding section for each of the topics listed at the beginning of
the file for consistency.

* CRAYSAT-1782: Add SAT-specific command-prompt conventions

There are a couple of SAT-specific command-prompt conventions that need
to be described in the CSM documentation command prompt conventions when
migrating docs-sat content into docs-csm.

The first is just the command-prompt convention for shells created by
the `sat bash` command. This is very similar to shells in Kubernetes
pods obtained by `kubectl exec`.

The second is a command-prompt convention for shells in which a Python
virtual environment is active. This is used in documentation that
describes installation and use of `sat` on external systems.

* CRAYSAT-1782: Migrate "About SAT" topics to docs-csm

Migrate the topics under the "About SAT" topic from docs-sat to
docs-csm. The "View SAT Documentation" topic has already been
generalized to CSM and moved into docs-csm in a previous commit, so that
topic did not migrate under the "About SAT" topic.

In addition, minor edits were made as follows:

* Update a link in `command_authentication.md` to link directly to the
  relevant markdown file since its now in the same repository.
* Reword the "SAT dependencies" topic slightly since now a lot of these
  dependencies are within the same CSM product.
* Remove command-prompt conventions from "Introduction to SAT" since
  the SAT-specific conventions have been migrated to docs-csm in a
  separate commit
* Move "SAT in CSM" topic under `about_sat` directory.
* Improve and update documentation in "Introduction to SAT" to improve
  clarity and examples surrounding man page viewing.
* Change file names to follow docs-csm conventions where file name
  matches top-level title in file.

* CRAYSAT-1782: Migrate SAT configuration and usage into docs-csm

Migrate the post-install configuration and usage topics from docs-sat
into docs-csm. The main changes from docs-sat are as follows:

* Use the CSM file naming convention where file names match top-level
  topics in the file.
* Change any topic references to CSM topics into actual links to the
  markdown files, since they are now in the same repo.
* Split the post-install configuration procedures into separate files.

* CRAYSAT-1782: Improve SAT authentication documentation

The documentation in files `about_sat/SAT_Command_Authentication.md` and
`configuration/Authenticate_SAT_Commands.md` overlapped in content.
Resolve this by changing "SAT Command Authentication" to be "SAT Command
Overview". The table in this file describes more than just the
authentication required for each command. Move all information about
actually running `sat auth` to perform authentication against the API
gateway to the configuration procedure "Authenticate SAT Commands".

* CRAYSAT-1782: Migrate "SAT on an External System" Documentation

Migrate the section "SAT on an External System" from docs-sat to
docs-csm. In doing so, make some minor edit, including:

* Remove some redundant explanation about authentication and instead
  refer to "Authenticate SAT Commands" section.
* Update `sat` release branch to 3.26 for a more current example in
  instructions.
* Change a link from HTML docs-csm to point directly at markdown file
  now that it is in the same repo.

* CRAYSAT-1782: Migrate uninstall/downgrade info from docs-sat

Migrate the topics on uninstalling and downgrading SAT versions from
docs-sat to docs-csm. Now that SAT is no longer a separate product and
is instead fully included in CSM v1.6.0, it can no longer be separately
uninstalled/removed by `prodmgr`. Update the documentation to note that
the uninstall instructions are only valid for the versions of SAT
installed as a separate product.

In addition, the ability to activate other versions of `sat` with
`prodmgr activate` is no longer supported. Instead, document how to more
simply switch between versions of the `cray-sat` container image using
the `SAT_IMAGE` environment variable. Move the documentation on this
procedure which was in the FAQ on the "SAT in CSM" topic into the
separate document which discusses uninstall and downgrade.

* CRAYSAT-1782: Reference SAT configuration in install procedure

Refer the reader to the appropriate SAT configuration procedures at the
appropriate time in the CSM install procedure. This should be done at
the same time as the `cray` CLI configuration and authentication.

This should not be necessary during upgrades as the authentication
tokens should remain in place, and I don't see anywhere it is re-done
for the `cray` CLI as part of the upgrade procedure.

* CRAYSAT-1782: Add spellchecker exceptions for SAT docs

There are a few spellchecker exceptions needed based on SAT
documentation content which has been migrated from docs-sat to docs-csm.

* CRAYSAT-1782: Change structure of "SAT Dependencies" topic

The markdownlint style checker complains about multiple headings with
the same name in a single file, even if they are not sibling headings.
docs-sat had limited the check on duplicate headings to only apply to
sibling headings. Since docs-csm has not limited that check to only
siblings, it was failing the style check. Replace the headings that
specified product names with just normal paragraphs that specify product
names.