CASMTRIAGE-6502 Handle multiple USB drives
The existing code fails if more than one USB drive is present. The code should be more tolerant, and permit any number of USB drives.

Ejection failure should not be fatal. This is script is a "best try" script for standing down the node to
facilitate faster/easier reboots. It isn't necessarily a script that
should flunk the entire run when USB ejection fails.