CASMTRIAGE-6606 CASMINST-6784 CASMINST-6785 CASMINST-6787 CSM 1.4.4 Fixes (#4681)
* CASMINST-6787 Fix invalid CFS session names

CFS session names are too long, preventing them from starting. CFS session names have a max of 45 characters for CSM 1.4. This change reduces our names to fit within the constraints.

* CASMTRIAGE-6606 Fix invalid path for `assign-ncn-images.sh`

`assign-ncn-images.sh` is not in a `configuration` subdirectory.

* CASMTRIAGE-6606 Fetch correct IMS image ID

The IMS image id for the base image was being fetched instead of the newly, built image. The new image was not getting assigned to the nodes, meaning the image rebuild was not propagating the 1.4.4 changes to SP4 NCNs.

* CASMTRIAGE-6606 Add a return

The hyperlink for returning to the 1.4.4 installation from the CSM only page was blank/missing.

* Ensure the `cephfs` share is mounted.

* CASMTRIAGE-6606 Ensure users change into the `cephfs` share

Before downloading the tar, users will want to change into the `cephfs` share. Without this step, none of the directions will align with the downloaded tar location.

* CASMINST-6785 CASMTRIAGE-6606 Incorrect variable

Remove `CSM_RELEASE_VERSION`, and use the more common variable of `CSM_RELEASE`. Also remove re-setting this variable, as users will set this when they download the tarball.

* Download latest docs as early as possible

* CASMTRIAGE-6606 Remind users of context

Remind users resuming at this stage that necessary variables were set earlier.

* CASMINST-6785 CASMTRIAGE-6606 Use `CSM_DISTDIR`

* CASMINST-6785 CASMTRIAGE-6606 `myenv`

Ensure `CSM_DISTDIR` is preserved

* CASMINST-6784 Ensure `cray-site-init` is installed

On starlord, `csi` was not installed on ncn-m001, this caused the rebuild procedure to fail out-of-the gate for masters.

* CASMTRIAGE-6606 Do not start another typescript

Continue using the typescript from the start of the 1.4.4 directions.