Fix short-circuit
This is a nit.

The pattern that exists here would successfull match any 1.4.5+ release,
however the pattern was a little sloppy in the sense that it wasn't
finding an exact match with some numbers.

- 1.4.50-1.4.99 were incomplete matches, albeit successful.
- Every first 10 digits of each hundred were also incomplete (1.4.202
  would match as 1.4.20)
- Every 10th number starting with 50 in every hundred were also
  incomplete (e.g., 1.4.500, 1.4.600)

This regex is just more complete, and less greedy. The problem was that
the regex group short-circuits, since `[5-9]` was first in the group
then the match was declared before checking the rest of the group.