CASMTRIAGE-7016: Fix update_tags.sh by removing bad `echo`
The `echo` command inside of `get_latest_tag_for_image` pollutes the
output of this function, which is then fed into `update_tags_in_file`,
which then results in the incorrect arguments being passed to a `sed`
command which results in `sed` failing with a "No such file or
directory" error.

Remove this `echo` command. It's not clear what its purpose is.

Test Description:
Removed this echo command from a copy of the script and executed the
script and verified that it correctly updated image tags in Argo
workflow templates.