CRAYSAT-1891: Correct inconsistencies in cabinet power documentation (#5299)
Referring to the mentioning in the ticket. It is identifed that the power_on_cabinets & power_off_cabinets
have been copied to the PCS folder. but no changes have been necessarily made to show the difference while using capmc & pcs.
Hence removing the duplicate files in the PCS folder and accorindgly updating the README.md file as well.