CRAYSAT-1909: Remove "verify Lustre access" step from power on procedure (#5417)
Remove the vague "verify Lustre access" step from the system power on
procedure. The step is too vague to be useful as it is written, and it
probably doesn't make sense to do this step at this stage in the system
power on procedure as the Kubernetes cluster isn't up yet, and the
Slingshot fabric will not be initialized yet.