CASMINST-7035: Improve "SAT and IUF" section
Improve the "SAT and IUF" section of the documentation that describes
the relationship between SAT and IUF, specifically how `sat
bootprep` is executed in IUF and how variable substitution works in
that context.

Include a procedure to obtain the `session_vars.yaml` file and
bootprep input files from the IUF activity and use them to create CFS
configurations, images, and session templates manually. This can be a
useful debugging procedure.

Include another procedure that describes how to assign the CFS
configurations and IMS images to management NCNs as a non-standard
debugging procedure. This procedure used to be used when performing an
upgrade of only CSM on a system that also contains other products that
were previously installed by IUF.

Move a section about limiting `sat bootprep run` into separate stages
into the `SAT_Bootprep.md` file because it's really more appropriate
with the general `sat bootprep` usage information.