CASMINST-6968 Fixes for 1.6 (#5639)
* Fix missing admission plugin

The list of plugins given to `enable-admission-plugins` was being written to the wrong top-level key. Instead of defining `.scheduler.extraArgs.enable-admission-plugins` we need to define `.apiServer.extraArgs.enable-admission-plugins`.

* Missing updated `.data.ClusterConfiguration`

The updates to `.data.ClusterConfiguration` were not saving into the `kubeadm-config.yaml` file. This change ensures `yq4` writes the updated config before we read it via `kubectl apply`.

* Fix false positive for enabling auditing

Translating the original code went ary during the auditing conditional. Now that both checks return `1` for true and `0` for false (instead of mix-n-matching) the conditional can treat `0` and `1` the same across both values.

Now `cm_auditing_enabled=1` means true, vs. before this meant false.

The conditional now only runs when both values are `1`.

* Fix `ImagePull` errors

The translation from `sed` to `awk` over-zealously set the DNS options for `kubeadm`, pointing to an unusable repository that resulting in `ImagePull` errors. According to the previous version of this script, these values were only set in certain cases and should not be defaults.