MTL-2126 Use boot order script
This replaces the long-standing manual directions with a script invocation. The manual directions for boot trimming were from the early days when we wanted to reproduce what automation was doing. This section has lost some of its value as a manual process, it should invoke the automation script for easier use.