CASMSEC-398: Docs: Port security hardening guide to 1.4, 1.5+ (#6005)
* Audit parameters enablement steps

This file was present in CSM 1.3 and removed from CSM 1.4. This is required for enabling audit log parameters. Hence reintroducing this in CSM 1.7. This also helps to adhere to some of the CIS benchmarks.

Signed-off-by: pradeepkumargl <80017325+pradeepkumargl@users.noreply.github.com>

* Required to achieve security hardening

This file contains the the links to different files which in tern have steps to achieve security hardening in CSM releases. This is removed post CSM 1.3 release. Reintroducing this as it is important to perform CSM hardening to comply with CIS.

Signed-off-by: pradeepkumargl <80017325+pradeepkumargl@users.noreply.github.com>

* Update Limit_Kubernetes_API_Audit_Log_Parameters.md fixing build error

Signed-off-by: pradeepkumargl <80017325+pradeepkumargl@users.noreply.github.com>

* Update Apply_Security_Hardening.md : Removed the audit enablement part

Based on Discussion with David removing the audit enablement part in Hardening guide to bring back the old file which is still relevant and needed for latest CSM release.

Signed-off-by: pradeepkumargl <80017325+pradeepkumargl@users.noreply.github.com>

* Delete operations/kubernetes/Limit_Kubernetes_API_Audit_Log_Parameters.md

This file is no longer relevant. As David suggested the audit parameters are set differently.

Signed-off-by: pradeepkumargl <80017325+pradeepkumargl@users.noreply.github.com>

---------

Signed-off-by: pradeepkumargl <80017325+pradeepkumargl@users.noreply.github.com>