CASMTRIAGE-8340: update platform-utils in prerequisites
Update prerequisites.sh to update platform-utils before running goss
tests. A particular etcd test requires an updated version of
platform-utils, which is normally not available until nodes reboot into
new images, but which we now need to have beforehand.

An updated platform-utils ought to be available in Nexus by this point,
so zypper should fetch the newest version and install it.