CASMTRIAGE-8340: update platform utils to 1.7.1 in prerequisites (#6088)
* CASMTRIAGE-8340: update platform-utils in prerequisites

Update prerequisites.sh to install platform-utils 1.7.1 before running
goss tests. A particular etcd test requires an updated version of
platform-utils, which is normally not available until nodes reboot into
new images, but which we now need to have beforehand. Additionally, we
need to use 1.7.1 instead of the latest version, because 1.8.0 and
higher require python11-boto3, which is unavailable in CSM 1.6, so we
use a version with the bugfix backported.

* Remove tabs from comment lines.