CASMTRIAGE-8373 drax: prerequisites.sh, UPDATE_CRAY_POSTGRES_OPERATOR_CRDS failing.
When upgrading drax to 1.7 via:
    1.3.5 -> 1.4.4 -> 1.5.4 -> 1.6.2-rc.1 -> 1.7.0-beta.1
the postgres CRD update in CSM 1.7 failed with
   The CustomResourceDefinition "operatorconfigurations.acid.zalan.do"
   is invalid: spec.preserveUnknownFields: Invalid value: true: must
   be false in order to use defaults in the schema
This does not happen on an upgrade from a fresh 1.6 install,
so this is carry-over from an earlier install or upgrade.

This is fixed by patching the current CRDs to remove the
spec.preserveUnknownFields field, prior to updating the CRDs.

And since we're patching the CRDs anyway, also remove the
"last-applied-configuration" annotation that kubectl added.